backend/requirements.txt
# These should be versioned but leaving unversioned for now
fastapi
uvicorn
torch == 1.8.1
torchvision == 0.9.1
numpy
Pillow
matplotlib
python-multipart
prometheus_fastapi_instrumentator